Improve windows flashing script
[flashing] Improve windows flashing script contributes to JB#39943

In some cases there are spaces in username and thus in the directory
path so to avoid that we need to get the filename from the full path.
Also the oem image is not included to the md5.lst file so we need to
skip md5sum checking for that file for now.

Signed-off-by: Marko Saukko <marko.saukko@jolla.com>